ref: move model into services folder
