+++ src/app/services/myinfo/myinfo_hash.model.ts
@@ -1,10 +1,9 @@
 import crypto from 'crypto'
 import { Mongoose, Schema } from 'mongoose'
 
-import { sessionSecret } from '../../config/config'
-import { IHashes, IMyInfoHashModel, IMyInfoHashSchema } from '../../types'
-
-import { FORM_SCHEMA_ID } from './form.server.model'
+import { sessionSecret } from '../../../config/config'
+import { IHashes, IMyInfoHashModel, IMyInfoHashSchema } from '../../../types'
+import { FORM_SCHEMA_ID } from '../../models/form.server.model'
 
 export const MYINFO_HASH_SCHEMA_ID = 'MyInfoHash'
